requirements.txt
altair==4.2.0
attrs==22.1.0
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
decorator==5.1.1
distlib==0.3.6
entrypoints==0.4
filelock==3.8.0
gitdb==4.0.9
GitPython==3.1.29
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
Jinja2==3.1.2
jsonschema==4.17.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pipenv==2022.10.25
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
protobuf==3.20.1
pyarrow==10.0.0
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==12.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.15.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
virtualenv==20.16.5
virtualenv-clone==0.5.7
watchdog==2.1.9
zipp==3.10.0
